Split test_blueprint.py
Send mixin tests to dedicated test file
